+++ app/src/main/java/test/com/jianyue/reader_activity/Read_activity/Bottom_list/FavoriteActivity.java
@@ -7,7 +7,13 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.ListView;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import test.com.jianyue.R;
+import test.com.jianyue.reader_activity.Read_activity.Articles;
+import test.com.jianyue.reader_activity.Read_activity.LikesAdapter;
+import test.com.jianyue.reader_activity.Read_activity.MainActivity;
 
 public class FavoriteActivity extends AppCompatActivity {
 
@@ -15,11 +21,16 @@ public class FavoriteActivity extends AppCompatActivity {
     public ActionBar actionBar;
     public ListView listView;
 
+    protected List<Articles> articlesList = new ArrayList<Articles>();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_favorite);
         listView=findViewById(R.id.listView);
+        initArticles();
+        LikesAdapter adapter = new LikesAdapter(FavoriteActivity.this, R.layout.article_item, articlesList);
+        listView.setAdapter(adapter);
         toolbar=findViewById(R.id.toolbar2);
         setSupportActionBar(toolbar);//toolbar绑定为actionbar
         actionBar = getSupportActionBar();//启用toolbar
@@ -34,5 +45,11 @@ public class FavoriteActivity extends AppCompatActivity {
                 finish();
             }
         });
+    }
+
+    //初始化收藏夹数据
+    private void initArticles() {
+        Articles meiwen = new Articles();
+        articlesList.add(meiwen);
     }
 }
